+++ packages/patternfly-3/patternfly-react/src/components/DualListSelector/DualList.stories.js
@@ -1,3 +1,4 @@
+/* eslint-disable jsx-a11y/iframe-has-title */
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
@@ -5,6 +6,7 @@ import { defaultTemplate } from 'storybook/decorators/storyTemplates';
 import { storybookPackageName, DOCUMENTATION_URL, STORYBOOK_CATEGORY } from 'storybook/constants/siteConstants';
 import { name } from '../../../package.json';
 import { items, dropdownItems } from './DualListMocks';
+import { Button } from '../../index';
 import {
   DualList,
   DualListArrows,
@@ -26,6 +28,24 @@ const stories = storiesOf(
   module
 );
 
+const storyInfo = {
+  source: false,
+  propTables: [
+    DualList,
+    DualListArrows,
+    DualListBody,
+    DualListCounter,
+    DualListFilter,
+    DualListFooter,
+    DualListHeading,
+    DualListItem,
+    DualListItems,
+    DualListMainCheckbox,
+    DualListSort
+  ],
+  propTablesExclude: [DualListControlled, DualListSelector]
+};
+
 stories.addDecorator(
   defaultTemplate({
     title: 'Dual List Selector',
@@ -35,32 +55,52 @@ stories.addDecorator(
 
 stories.add(
   'Dual List Selector',
-  withInfo({
-    source: false,
-    propTables: [
-      DualList,
-      DualListArrows,
-      DualListBody,
-      DualListCounter,
-      DualListFilter,
-      DualListFooter,
-      DualListHeading,
-      DualListItem,
-      DualListItems,
-      DualListMainCheckbox,
-      DualListSort
-    ],
-    propTablesExclude: [DualListControlled, DualListSelector]
-  })(() => (
+  withInfo(storyInfo)(() => (
     <DualListControlled
       left={{
         items: items.left,
+        inputProps: { name: 'left' },
         kebabMenu: dropdownItems
       }}
       right={{
         items: items.right,
+        inputProps: { name: 'right' },
         kebabMenu: dropdownItems
       }}
     />
+  ))
+);
+
+stories.add(
+  'Dual List Selector as a form input',
+  withInfo(storyInfo)(() => (
+    <React.Fragment>
+      <h3>Please see this short video to understand how to send the form values by network:</h3>
+      <iframe
+        width="400"
+        height="200"
+        style={{ marginBottom: '20px' }}
+        src="https://www.youtube.com/embed/G8rlyyifhtQ"
+        frameBorder="0"
+        allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
+        allowFullScreen
+      />
+      <form action="">
+        <DualListControlled
+          allowHiddenInputs
+          left={{
+            items: items.left,
+            inputProps: { name: 'left-selector' },
+            kebabMenu: dropdownItems
+          }}
+          right={{
+            items: items.right,
+            inputProps: { name: 'right-selector' },
+            kebabMenu: dropdownItems
+          }}
+        />
+        <Button type="submit">Submit form</Button>
+      </form>
+    </React.Fragment>
   ))
 );