+++ packages/patternfly-4/react-core/src/components/Avatar/index.ts
@@ -1,0 +1,1 @@
+export { default as Avatar, AvatarProps } from './Avatar';